# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlayMoney API",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "pageSize": {
        "type": [
          "number",
          "null"
        ]
      },
      "status": {
        "type": "string",
        "enum": [
          "active",
          "closed",
          "all"
        ]
      },
...
      "page": {
        "type": [
          "number",
          "null"
        ]
      },
...
      "createdBy": {
        "type": "string"
      },
      "tag": {
        "type": "string"
      },
      "sortField": {
        "type": "string"
      },
      "sortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
    },
    "parameters": {
...
      "pageSize": {
        "schema": {
          "$ref": "#/components/schemas/pageSize"
        },
        "required": false,
        "name": "pageSize",
        "in": "path"
      },
      "status": {
        "schema": {
          "$ref": "#/components/schemas/status"
        },
        "required": false,
        "name": "status",
        "in": "path"
      },
...
      "page": {
        "schema": {
          "$ref": "#/components/schemas/page"
        },
        "required": false,
        "name": "page",
        "in": "path"
      },
...
      "createdBy": {
        "schema": {
          "$ref": "#/components/schemas/createdBy"
        },
        "required": false,
        "name": "createdBy",
        "in": "path"
      },
      "tag": {
        "schema": {
          "$ref": "#/components/schemas/tag"
        },
        "required": false,
        "name": "tag",
        "in": "path"
      },
      "sortField": {
        "schema": {
          "$ref": "#/components/schemas/sortField"
        },
        "required": false,
        "name": "sortField",
        "in": "path"
      },
      "sortDirection": {
        "schema": {
          "$ref": "#/components/schemas/sortDirection"
        },
        "required": false,
        "name": "sortDirection",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/markets": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/createdBy"
          },
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/tag"
          },
          {
            "$ref": "#/components/parameters/sortField"
          },
          {
            "$ref": "#/components/parameters/sortDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "markets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "cuid"
                          },
                          "question": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string",
                            "minLength": 1
                          },
                          "closeDate": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "resolvedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "canceledAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "canceledById": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "maxItems": 5
                          },
                          "ammAccountId": {
                            "type": "string"
                          },
                          "clearingAccountId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "updatedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "commentCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "uniqueTradersCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "uniquePromotersCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "liquidityCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "parentListId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "question",
                          "description",
                          "slug",
                          "closeDate",
                          "resolvedAt",
                          "canceledAt",
                          "canceledById",
                          "createdBy",
                          "tags",
                          "ammAccountId",
                          "clearingAccountId",
                          "createdAt",
                          "updatedAt",
                          "commentCount",
                          "uniqueTradersCount",
                          "uniquePromotersCount",
                          "liquidityCount",
                          "parentListId"
                        ]
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "markets",
                    "page",
                    "pageSize",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "closeDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 5
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "color": {
                          "type": "string",
                          "pattern": "^#[0-9A-Fa-f]{6}$"
                        }
                      },
                      "required": [
                        "name",
                        "color"
                      ]
                    }
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "binary",
                      "multi",
                      "list"
                    ]
                  },
                  "contributionPolicy": {
                    "type": "string",
                    "enum": [
                      "DISABLED",
                      "OWNERS_ONLY",
                      "FRIENDS_ONLY",
                      "PUBLIC"
                    ]
                  }
                },
                "required": [
                  "question",
                  "description",
                  "closeDate",
                  "tags",
                  "options"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "market": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid"
                        },
                        "question": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "closeDate": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "resolvedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "canceledAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "canceledById": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "createdBy": {
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "maxItems": 5
                        },
                        "ammAccountId": {
                          "type": "string"
                        },
                        "clearingAccountId": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "updatedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "commentCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "uniqueTradersCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "uniquePromotersCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "liquidityCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "parentListId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "question",
                        "description",
                        "slug",
                        "closeDate",
                        "resolvedAt",
                        "canceledAt",
                        "canceledById",
                        "createdBy",
                        "tags",
                        "ammAccountId",
                        "clearingAccountId",
                        "createdAt",
                        "updatedAt",
                        "commentCount",
                        "uniqueTradersCount",
                        "uniquePromotersCount",
                        "liquidityCount",
                        "parentListId"
                      ]
                    },
                    "list": {
                      "type": "object",
                      "properties": {
                        "contributionPolicy": {
                          "type": "string",
                          "enum": [
                            "DISABLED",
                            "OWNERS_ONLY",
                            "FRIENDS_ONLY",
                            "PUBLIC"
                          ]
                        },
                        "id": {
                          "type": "string",
                          "format": "cuid"
                        },
                        "title": {
                          "type": "string",
                          "minLength": 1
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ownerId": {
                          "type": "string"
                        },
                        "contributionReview": {
                          "type": [
                            "boolean",
                            "null"
                          ]
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "maxItems": 5
                        },
                        "createdAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "updatedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "contributionPolicy",
                        "id",
                        "title",
                        "slug",
                        "description",
                        "ownerId",
                        "contributionReview",
                        "tags",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}